Refactor: Split analytics cli.py into specialized modules (#5450)
