+++ surfaces/cli/wizard/app.py
@@ -5,7 +5,8 @@ from __future__ import annotations
 import click
 
 from config.local_env import bootstrap_opensre_env_once
-from infrastructure.analytics.cli import build_cli_invoked_properties, capture_cli_invoked
+from infrastructure.analytics.event_properties import build_cli_invoked_properties
+from infrastructure.analytics.capture import capture_cli_invoked
 from infrastructure.analytics.provider import capture_first_run_if_needed, shutdown_analytics
 from infrastructure.observability.errors.sentry import init_sentry
 from infrastructure.terminal.prompt_support import install_questionary_escape_cancel
