+++ surfaces/shared/terminal/stream_renderer/diagnose.py
@@ -14,7 +14,7 @@ from rich.spinner import Spinner
 from rich.text import Text
 
 from core.domain.stream import StreamEvent
-from infrastructure.analytics.cli import capture_investigation_lifecycle_event
+from infrastructure.analytics.capture import capture_investigation_lifecycle_event
 from infrastructure.analytics.events import Event
 from surfaces.shared.terminal.output import (
     ProgressTracker,
